Separate `QuestList` to another file
