+++ Lib9c/Model/Quest/Quest.cs
@@ -1,12 +1,7 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
 using Bencodex.Types;
-using Libplanet.Assets;
-using Nekoyume.Model.EnumType;
-using Nekoyume.Model.Item;
 using Nekoyume.Model.State;
 using Nekoyume.TableData;
 using Serilog;
@@ -135,298 +130,6 @@ namespace Nekoyume.Model.Quest
         public static Quest Deserialize(IValue arg)
         {
             return Deserialize((Dictionary) arg);
-        }
-    }
-
-    [Serializable]
-    public class QuestList : IEnumerable<Quest>, IState
-    {
-        private readonly List<Quest> _quests;
-
-        private int _listVersion = 1;
-        
-        public List<int> completedQuestIds = new List<int>();
-
-        public QuestList(QuestSheet questSheet,
-            QuestRewardSheet questRewardSheet,
-            QuestItemRewardSheet questItemRewardSheet,
-            EquipmentItemRecipeSheet equipmentItemRecipeSheet,
-            EquipmentItemSubRecipeSheet equipmentItemSubRecipeSheet
-        )
-        {
-            _quests = new List<Quest>();
-            foreach (var questData in questSheet.OrderedList)
-            {
-                Quest quest;
-                QuestReward reward = GetQuestReward(
-                    questData.QuestRewardId,
-                    questRewardSheet,
-                    questItemRewardSheet
-                );
-                switch (questData)
-                {
-                    case CollectQuestSheet.Row row:
-                        quest = new CollectQuest(row, reward);
-                        _quests.Add(quest);
-                        break;
-                    case CombinationQuestSheet.Row row1:
-                        quest = new CombinationQuest(row1, reward);
-                        _quests.Add(quest);
-                        break;
-                    case GeneralQuestSheet.Row row2:
-                        quest = new GeneralQuest(row2, reward);
-                        _quests.Add(quest);
-                        break;
-                    case ItemEnhancementQuestSheet.Row row3:
-                        quest = new ItemEnhancementQuest(row3, reward);
-                        _quests.Add(quest);
-                        break;
-                    case ItemGradeQuestSheet.Row row4:
-                        quest = new ItemGradeQuest(row4, reward);
-                        _quests.Add(quest);
-                        break;
-                    case MonsterQuestSheet.Row row5:
-                        quest = new MonsterQuest(row5, reward);
-                        _quests.Add(quest);
-                        break;
-                    case TradeQuestSheet.Row row6:
-                        quest = new TradeQuest(row6, reward);
-                        _quests.Add(quest);
-                        break;
-                    case WorldQuestSheet.Row row7:
-                        quest = new WorldQuest(row7, reward);
-                        _quests.Add(quest);
-                        break;
-                    case ItemTypeCollectQuestSheet.Row row8:
-                        quest = new ItemTypeCollectQuest(row8, reward);
-                        _quests.Add(quest);
-                        break;
-                    case GoldQuestSheet.Row row9:
-                        quest = new GoldQuest(row9, reward);
-                        _quests.Add(quest);
-                        break;
-                    case CombinationEquipmentQuestSheet.Row row10:
-                        int stageId;
-                        var recipeRow = equipmentItemRecipeSheet.Values
-                            .FirstOrDefault(r => r.Id == row10.RecipeId);
-                        if (recipeRow is null)
-                        {
-                            throw new ArgumentException($"Invalid Recipe Id : {row10.RecipeId}");
-                        }
-
-                        stageId = recipeRow.UnlockStage;
-                        quest = new CombinationEquipmentQuest(row10, reward, stageId);
-                        _quests.Add(quest);
-                        break;
-                }
-            }
-        }
-
-
-        public QuestList(Dictionary serialized)
-        {
-            _listVersion = serialized.TryGetValue((Text) "list-version", out var listVersion)
-                ? listVersion.ToInteger()
-                : 1;
-            
-            _quests = serialized.TryGetValue((Text) "quests", out var questsValue)
-                ? questsValue.ToList(Quest.Deserialize)
-                : new List<Quest>();
-
-            completedQuestIds = serialized.TryGetValue((Text) "completedQuestIds", out var idsValue)
-                ? idsValue.ToList(StateExtensions.ToInteger)
-                : new List<int>();
-        }
-
-        public IEnumerator<Quest> GetEnumerator()
-        {
-            return _quests.OrderBy(q => q.Id).GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-
-        public void UpdateCombinationQuest(ItemUsable itemUsable)
-        {
-            var targets = _quests
-                .OfType<CombinationQuest>()
-                .Where(i => i.ItemType == itemUsable.ItemType &&
-                            i.ItemSubType == itemUsable.ItemSubType &&
-                            !i.Complete);
-            foreach (var target in targets)
-            {
-                target.Update(new List<ItemBase> {itemUsable});
-            }
-        }
-
-        public void UpdateTradeQuest(TradeType type, FungibleAssetValue price)
-        {
-            var tradeQuests = _quests
-                .OfType<TradeQuest>()
-                .Where(i => i.Type == type && !i.Complete);
-            foreach (var tradeQuest in tradeQuests)
-            {
-                tradeQuest.Check();
-            }
-
-            var goldQuests = _quests
-                .OfType<GoldQuest>()
-                .Where(i => i.Type == type && !i.Complete);
-            foreach (var goldQuest in goldQuests)
-            {
-                goldQuest.Update(price);
-            }
-        }
-
-        public void UpdateStageQuest(CollectionMap stageMap)
-        {
-            var stageQuests = _quests.OfType<WorldQuest>();
-            foreach (var quest in stageQuests)
-            {
-                quest.Update(stageMap);
-            }
-        }
-
-        public void UpdateMonsterQuest(CollectionMap monsterMap)
-        {
-            var monsterQuests = _quests.OfType<MonsterQuest>();
-            foreach (var quest in monsterQuests)
-            {
-                quest.Update(monsterMap);
-            }
-        }
-
-        public void UpdateCollectQuest(CollectionMap itemMap)
-        {
-            var collectQuests = _quests.OfType<CollectQuest>();
-            foreach (var quest in collectQuests)
-            {
-                quest.Update(itemMap);
-            }
-        }
-
-        public void UpdateItemEnhancementQuest(Equipment equipment)
-        {
-            var targets = _quests
-                .OfType<ItemEnhancementQuest>()
-                .Where(i => !i.Complete && i.Grade == equipment.Grade);
-            foreach (var target in targets)
-            {
-                target.Update(equipment);
-            }
-        }
-
-        public CollectionMap UpdateGeneralQuest(IEnumerable<QuestEventType> types,
-            CollectionMap eventMap)
-        {
-            foreach (var type in types)
-            {
-                var targets = _quests
-                    .OfType<GeneralQuest>()
-                    .Where(i => i.Event == type && !i.Complete);
-                foreach (var target in targets)
-                {
-                    target.Update(eventMap);
-                }
-            }
-
-            return eventMap;
-        }
-
-        public void UpdateItemGradeQuest(ItemUsable itemUsable)
-        {
-            var targets = _quests
-                .OfType<ItemGradeQuest>()
-                .Where(i => i.Grade == itemUsable.Grade && !i.Complete);
-            foreach (var target in targets)
-            {
-                target.Update(itemUsable);
-            }
-        }
-
-        public void UpdateItemTypeCollectQuest(IEnumerable<ItemBase> items)
-        {
-            if (items == null)
-            {
-                return;
-            }
-            
-            foreach (var item in items.OrderBy(i => i.Id))
-            {
-                var targets = _quests
-                    .OfType<ItemTypeCollectQuest>()
-                    .Where(i => i.ItemType == item.ItemType && !i.Complete);
-                foreach (var target in targets)
-                {
-                    target.Update(item);
-                }
-            }
-        }
-
-        public IValue Serialize()
-        {
-            if (_listVersion > 1)
-            {
-                return Dictionary.Empty
-                    .SetItem("list-version", _listVersion.Serialize())
-                    .SetItem("quests", (IValue) new List(_quests
-                        .OrderBy(i => i.Id)
-                        .Select(q => q.Serialize())))
-                    .SetItem("completedQuestIds", (IValue) new List(completedQuestIds
-                        .OrderBy(i => i)
-                        .Select(i => i.Serialize())));
-            }
-
-            return new Dictionary(new Dictionary<IKey, IValue>
-            {
-                [(Text) "quests"] = new List(_quests
-                    .OrderBy(i => i.Id)
-                    .Select(q => q.Serialize())),
-                [(Text) "completedQuestIds"] = new List(completedQuestIds
-                    .OrderBy(i => i)
-                    .Select(i => i.Serialize()))
-            });
-        }
-
-        public void UpdateCombinationEquipmentQuest(int recipeId)
-        {
-            var targets = _quests.OfType<CombinationEquipmentQuest>()
-                .Where(q => !q.Complete);
-            foreach (var target in targets)
-            {
-                target.Update(recipeId);
-            }
-        }
-
-
-        public CollectionMap UpdateCompletedQuest(CollectionMap eventMap)
-        {
-            const QuestEventType type = QuestEventType.Complete;
-            eventMap[(int) type] = _quests.Count(i => i.Complete);
-            return UpdateGeneralQuest(new[] {type}, eventMap);
-        }
-
-
-        private static QuestReward GetQuestReward(
-            int rewardId,
-            QuestRewardSheet rewardSheet,
-            QuestItemRewardSheet itemRewardSheet)
-        {
-            var itemMap = new Dictionary<int, int>();
-            if (rewardSheet.TryGetValue(rewardId, out var questRewardRow))
-            {
-                foreach (var id in questRewardRow.RewardIds.OrderBy(i => i))
-                {
-                    if (itemRewardSheet.TryGetValue(id, out var itemRewardRow))
-                    {
-                        itemMap[itemRewardRow.ItemId] = itemRewardRow.Count;
-                    }
-                }
-            }
-
-            return new QuestReward(itemMap);
         }
     }
 }
